Rename useObserver file to use_observer.dart
