+++ flutter_mobx_hooks/lib/flutter_mobx_hooks.dart
@@ -1,3 +1,3 @@
 library flutter_mobx_hooks;
 
-export 'package:flutter_mobx_hooks/src/framework.dart' show useObserver;
+export 'package:flutter_mobx_hooks/src/use_observer.dart' show useObserver;
