+++ flutter_mobx_hooks/test/all.dart
@@ -1,0 +1,5 @@
+import './use_observer_test.dart' as use_observer_test;
+
+void main() {
+  use_observer_test.main();
+}
